Remove lightning owner/controller detachment
There was never any scenarios where owner and controller would MATTER to
be different. The original shooter(owner) didn't matter in the end
(damagesource couldn't reference it) so this feature was essentially
useless.

(DamageSource can only keep track of 2 entities; the target and
attacker. Couldn't have a redirection DamageSource that keeps track of
original shooter, redirector, and target)