Fix PD packet not working b/c reading ByteBuf deallocated (fixes #40)
In PacketCPlayerData reading, previously it held onto the ByteBuf and used
it later to read the data. However, a change in Forge code now makes this
bytebuf get deallocated and unreadable.

This made client not receive any information about bending etc. Therefore
bending was apparently disabled.

To fix, just make a copy of the bytebuf