Fix performance always added even when unsuccessful attack
Previously, performance was added every tick that a projectile collided
with an enemy. However, this resulted in "spamming" the performance boost,
since projectiles usually collide with enemies for multiple ticks (before
the enemy is pushed sufficient distance away to not collide anymore).

To avoid this, and only add performance ONCE, now projectiles will only
add performance when it is a SUCCESSFUL hit, i.e. they actually damaged
the enemy. This works because mobs have a "damage resistance" time for
0.5s after being attacked. This is used here as a way to make sure that they are only adding performance
once.