Use explicitly configured chi amounts for buff abilities
Previously buff abilities just used the default amount of chi, then
multiplied it by a number depending on the level. This was a problem
because if the multiplier number was big enough, it could be too big for
the player's chi capacity. Now, the buff ability uses a value explicitly
in the configuration for each level